tests/wpt/web-platform-tests/html/canvas/element/fill-and-stroke-styles/2d.pattern.modify.image2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.pattern.modify.image2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.pattern.modify.image2</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

var img = document.getElementById('green.png');
var pattern = ctx.createPattern(img, 'no-repeat');
ctx.fillStyle = pattern;
ctx.fillRect(0, 0, 100, 50);
ctx.fillStyle = '#00f';
ctx.fillRect(0, 0, 100, 50);
deferTest();
img.onload = t.step_func_done(function ()
{
    ctx.fillStyle = pattern;
    ctx.fillRect(0, 0, 100, 50);

    _assertPixel(canvas, 1,1, 0,255,0,255, "1,1", "0,255,0,255");
    _assertPixel(canvas, 98,1, 0,255,0,255, "98,1", "0,255,0,255");
    _assertPixel(canvas, 1,48, 0,255,0,255, "1,48", "0,255,0,255");
    _assertPixel(canvas, 98,48, 0,255,0,255, "98,48", "0,255,0,255");
});
img.src = '/images/red.png';


});
</script>
<img src="/images/green.png" id="green.png" class="resource">
